pyproject.toml
[tool.poetry]
name = "my-llm-apps"
version = "0.1.0"
description = ""
authors = ["Junya Nakazato <nakazato_junya@ca-adv.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slack-bolt = "^1.18.1"
load-dotenv = "^0.1.0"
openai = "^1.12.0"
langchain = "^0.1.8"
langchain-openai = "^0.0.6"
slack-sdk = "^3.27.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0"
black = "^24.2"
isort = "^5.13"
flake8-isort = "^6.1"
flake8-bugbear = "^24.2"
flake8-builtins = "^2.2"
flake8-eradicate = "^1.5"
flake8-unused-arguments = "^0.0"
flake8-pytest-style = "^1.7"
pep8-naming = "^0.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 160

[tool.black]
target-version = ['py39']
line-length = 160

[tool.flake8]
max-line-length = 160
exclude = [".venv", ".git", "__pycache__"]
extend-ignore = ["E203", "W503"]
max-complexity = 18
